dreport: Added PEL to the BMC dump header
Change:
The change involves adding PEL information to the dump header
for checkstop and elog dump types in the dreport utility.
After implementing the changes, the PEL ID is successfully
included in the dump header.

Tested:
The changes have been tested, and the PEL ID is now properly
added to the dump header as intended.

Signed-off-by: Chirag Sharma <chirshar@in.ibm.com>