Add support for Memory Buffer SBE dump
The memory buffer SBE dump is similar to the dump collected
from the processor SBE, but with a different method.
This dump is collected during a critical failure on
the SBE and contain specific debug data.

This commit provides an interface calling the dump collection
so other services can invoke dump collection when there
is an SBE failure detected.

Signed-off-by: Dhruvaraj Subhashchandran <dhruvaraj@in.ibm.com>