Fix handling of dump status on collection or packaging failure
This commit addresses an issue where dump entries erroneously remained
in the "InProgress" state when errors occurred during the dump
collection or packaging processes. The absence of a status update
to "Failed" in such scenarios misled callers about the actual state
of the dump.

Tested:
The status got updated when the dump collection failed.

Signed-off-by: Dhruvaraj Subhashchandran <dhruvaraj@in.ibm.com>